Moved the struct implementation to the canonical one, fixed the array implementation
