+++ hail/src/main/scala/is/hail/types/physical/PBaseStruct.scala
@@ -189,16 +189,6 @@ abstract class PBaseStruct extends PType {
     } else
       Gen.uniformSequence(types.map(t => t.genValue)).map(a => Annotation(a: _*))
   }
-
-  override def unstagedStoreJavaObjectAtAddress(addr: Long, annotation: Annotation, region: Region): Unit = {
-    initialize(addr)
-    val row = annotation.asInstanceOf[Row]
-    // TODO Separate unsafe row handling
-    this.types.zipWithIndex.foreach { case (fieldPt, fieldIdx) =>
-      val fieldAddress = fieldOffset(addr, fieldIdx)
-      fieldPt.unstagedStoreJavaObjectAtAddress(fieldAddress, row(fieldIdx), region)
-    }
-  }
 }
 
 abstract class PBaseStructValue extends PValue with SBaseStructValue {
