+++ hail/src/main/scala/is/hail/types/physical/PCanonicalArray.scala
@@ -505,4 +505,21 @@ final case class PCanonicalArray(elementType: PType, required: Boolean = false) 
   }
 
   def loadFromNested(cb: EmitCodeBuilder, addr: Code[Long]): Code[Long] = Region.loadAddress(addr)
+
+  override def unstagedStoreJavaObjectAtAddress(addr: Long, annotation: Annotation, region: Region): Unit = {
+    // TODO Special case UnsafeIndexedSeq
+    val is = annotation.asInstanceOf[IndexedSeq[Annotation]]
+    val valueAddress = allocate(region, is.length)
+    assert(is.length > 0)
+    // TODO: Don't have to touch missing? Or do I?
+    initialize(valueAddress, is.length)
+    var i = 0
+    var curElementAddress = firstElementOffset(valueAddress, is.length)
+    while (i < is.length) {
+      elementType.unstagedStoreJavaObjectAtAddress(curElementAddress, is(i), region)
+      curElementAddress = nextElementAddress(curElementAddress)
+      i += 1
+    }
+    Region.storeAddress(addr, valueAddress)
+  }
 }
